cleaning up a bit and readding http proxy and pooledlifetime functionality
