+++ BlizzardApiReader.Core/ApiReader.cs
@@ -5,6 +5,9 @@ using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 using BlizzardApiReader.Core.Exceptions;
 using Microsoft.Extensions.Options;
+using Microsoft.Extensions.DependencyInjection;
+using BlizzardApiReader.Core.Enums;
+using System.Net.Http.Headers;
 
 namespace BlizzardApiReader.Core
 {
@@ -23,18 +26,15 @@ namespace BlizzardApiReader.Core
             get { return apiConfiguration ?? defaultConfig; }
             set
             {
-                apiConfiguration = value;
-                if (_webClient != null)
-                {
-                    _webClient.Initialize(Configuration);
-                }
+                apiConfiguration = value;               
             }
         }
 
         private readonly IWebClient _webClient;
+        
         private string _token;
         private DateTime _tokenExpiration;
-
+               
         public ApiReader()
         {
 
@@ -44,14 +44,14 @@ namespace BlizzardApiReader.Core
         {
             _webClient = webClient;
             
-            Configuration = apiConfiguration.Value;
+            Configuration = apiConfiguration.Value;                        
         }
 
         public static void SetDefaultConfiguration(ApiConfiguration configuration)
         {
             defaultConfig = configuration;
         }
-
+                
         public static void ClearDefaultConfiguration()
         {
             defaultConfig = null;
