+++ BlizzardApiReader.Core/DependencyInjection/IServiceCollectionExtension.cs
@@ -5,18 +5,40 @@ using BlizzardApiReader.Core.Enums;
 using System.Net.Http.Headers;
 using System.Linq;
 using System.Net.Http;
+using Microsoft.Extensions.Options;
 
 namespace BlizzardApiReader.Core.DependencyInjection
 {
     public static class IServiceCollectionExtension
     {        
-        public static IServiceCollection AddBlizzardApiReader(this IServiceCollection services)
+        public static IServiceCollection AddBlizzardApiReader(this IServiceCollection services, HttpConfiguration httpConfiguration =null )
         {
-                        
-            if (!services.Any(x => x.ServiceType == typeof(IHttpClientFactory)))
+            if (httpConfiguration != null)
             {
-                services.AddHttpClient(NamedHttpClients.ApiClient.ToString()).ConfigureHttpClient(client => client.DefaultRequestHeaders.Accept.Add(new MediaTypeWithQualityHeaderValue("application/json")));
-                services.AddHttpClient(NamedHttpClients.AuthClient.ToString()).ConfigureHttpClient(client => client.DefaultRequestHeaders.Accept.Add(new MediaTypeWithQualityHeaderValue("application/json")));
+                SocketsHttpHandler socketsHttpHandler;
+
+                socketsHttpHandler = new SocketsHttpHandler()
+                {
+                    Proxy = httpConfiguration.GetHttpProxy(),
+                    PooledConnectionLifetime = httpConfiguration.GetPooledConnectionLifetime()
+                };
+
+
+                services.AddHttpClient(NamedHttpClients.ApiClient.ToString())
+                    .ConfigurePrimaryHttpMessageHandler(() => socketsHttpHandler)
+                    .ConfigureHttpClient(client => client.DefaultRequestHeaders.Accept.Add(new MediaTypeWithQualityHeaderValue("application/json")));
+
+                services.AddHttpClient(NamedHttpClients.AuthClient.ToString())
+                    .ConfigurePrimaryHttpMessageHandler(() => socketsHttpHandler)
+                    .ConfigureHttpClient(client => client.DefaultRequestHeaders.Accept.Add(new MediaTypeWithQualityHeaderValue("application/json")));
+            }
+            else
+            {
+                services.AddHttpClient(NamedHttpClients.ApiClient.ToString())                    
+                    .ConfigureHttpClient(client => client.DefaultRequestHeaders.Accept.Add(new MediaTypeWithQualityHeaderValue("application/json")));
+
+                services.AddHttpClient(NamedHttpClients.AuthClient.ToString())                    
+                    .ConfigureHttpClient(client => client.DefaultRequestHeaders.Accept.Add(new MediaTypeWithQualityHeaderValue("application/json")));
             }
 
             //Using TryAdd... in case it is already added
@@ -25,7 +47,6 @@ namespace BlizzardApiReader.Core.DependencyInjection
             services.TryAddSingleton<IWebClient,ApiWebClient>();
                         
             return services;
-        }
-
+        }        
     }
 }
