+++ BlizzardApiReader.Core/Models/ApiConfiguration.cs
@@ -16,10 +16,7 @@ namespace BlizzardApiReader.Core
         private const string API_URL_TEMPLATE = "https://REGION.api.blizzard.com";
         private string authUrl = string.Empty;
         private string apiUrl = string.Empty;
-
-        private WebProxy _WebProxy = null;
-        private TimeSpan _PooledConnectionLifetime = TimeSpan.FromMinutes(1);
-
+                
         public ApiConfiguration()
         {
             SetRegion(Region.Europe, true);
@@ -72,41 +69,7 @@ namespace BlizzardApiReader.Core
             ResultLocale = locale;
             return this;
         }
-
-        /// <summary>
-        /// Set the http proxy to use on all the calls, by default is not configured (null) and to remove it you need to call it to with ""
-        /// </summary>
-        /// <param name="proxy">URL with the proxy to use, for example http://server:port </param>        
-        /// <returns>This instance of ApiConfiguration</returns>
-        public ApiConfiguration SetHttpProxy(string proxy)
-        {
-            if (proxy == "")
-            {
-                _WebProxy = null;
-            }
-            else
-            {
-                _WebProxy = new WebProxy()
-                {
-                    Address = new Uri(proxy),
-                    UseDefaultCredentials = true
-                };
-            }
-
-            return this;
-        }
-
-        /// <summary>
-        /// Configure the lifetime of the connection for refreshing DNS queries, by default is 1 minute
-        /// </summary>
-        /// <param name="pooledConnectionLifetime">Lifetime in minutes</param>        
-        /// <returns>This instance of ApiConfiguration</returns>
-        public ApiConfiguration SetPooledConnectionLifetime(int pooledConnectionLifetime)
-        {
-            _PooledConnectionLifetime = TimeSpan.FromMinutes(pooledConnectionLifetime);
-            return this;
-        }
-
+        
         /// <summary>
         /// Declare this Configuration as the global default configuration, it will be used when no configuration is provided to the api reader.
         /// </summary>  
@@ -134,15 +97,6 @@ namespace BlizzardApiReader.Core
         public string GetApiUrl()
         {
             return apiUrl;
-        }
-
-        public WebProxy GetHttpProxy()
-        {
-            return _WebProxy;
-        }
-        public TimeSpan GetPooledConnectionLifetime()
-        {
-            return _PooledConnectionLifetime;
         }
     }
 }
