+++ BlizzardApiReader.Core/Models/ApiWebClient.cs
@@ -1,4 +1,5 @@
 ﻿using BlizzardApiReader.Core.Enums;
+using Microsoft.Extensions.Options;
 using System;
 using System.Collections.Generic;
 using System.Net;
@@ -16,9 +17,17 @@ namespace BlizzardApiReader.Core
 
         private readonly IHttpClientFactory _httpClientFactory;
 
-        public ApiWebClient(IHttpClientFactory httpClientFactory)
+        public ApiWebClient(IHttpClientFactory httpClientFactory, IOptions<ApiConfiguration> apiConfiguration )
         {
+            var _configuration = apiConfiguration.Value;
+
             _httpClientFactory = httpClientFactory;
+
+            authPath = _configuration.GetAuthUrl();            
+
+            _apiClient = configureApiClient();
+            _authClient = configureAuthClient(authenticateHeader(_configuration.ClientId, _configuration.ClientSecret));
+
         }
 
 
@@ -29,17 +38,7 @@ namespace BlizzardApiReader.Core
                 });
 
         private string authPath;
-        private SocketsHttpHandler socketsHttpHandler;
-
-        public void Initialize(ApiConfiguration _configuration)
-        {
-            authPath = _configuration.GetAuthUrl();
-            socketsHttpHandler = createSocketsHandler(_configuration);
-
-            _apiClient = configureApiClient();
-            _authClient = configureAuthClient(authenticateHeader(_configuration.ClientId, _configuration.ClientSecret));
-        }
-
+                      
         public async Task<IApiResponse> MakeApiRequestAsync(string path)
         {
             var response = await _apiClient.GetAsync(path);
@@ -58,31 +57,7 @@ namespace BlizzardApiReader.Core
             var response = await _authClient.PostAsync(authPath, _authRequestContent);
             return new ApiResponse(response);
         }
-
-
-        private SocketsHttpHandler createSocketsHandler(ApiConfiguration _configuration)
-        {
-            SocketsHttpHandler socketsHttpHandler;
-            //if a proxy was configured use it for the connection
-            if (_configuration.GetHttpProxy() != null)
-            {
-                socketsHttpHandler = new SocketsHttpHandler()
-                {
-                    Proxy = _configuration.GetHttpProxy(),
-                    PooledConnectionLifetime = _configuration.GetPooledConnectionLifetime()
-                };
-            }
-            else
-            {
-                socketsHttpHandler = new SocketsHttpHandler()
-                {
-                    PooledConnectionLifetime = _configuration.GetPooledConnectionLifetime()
-                };
-            }
-
-            return socketsHttpHandler;
-        }
-
+        
         private HttpClient configureApiClient()
         {            
             var client = _httpClientFactory.CreateClient(NamedHttpClients.ApiClient.ToString());
