+++ BlizzardApiReader.Core/Models/IWebClient.cs
@@ -5,7 +5,6 @@ namespace BlizzardApiReader.Core
     public interface IWebClient
     {
         Task<IApiResponse> MakeApiRequestAsync(string path);
-        Task<IApiResponse> RequestAccessTokenAsync();
-        void Initialize(ApiConfiguration configuration);
+        Task<IApiResponse> RequestAccessTokenAsync();        
     }
 }
