+++ BlizzardApiReader.Diablo/DependencyInjection/IServiceCollectionExtension.cs
@@ -1,4 +1,5 @@
-﻿using BlizzardApiReader.Core.DependencyInjection;
+﻿using BlizzardApiReader.Core;
+using BlizzardApiReader.Core.DependencyInjection;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using System;
@@ -10,9 +11,9 @@ namespace BlizzardApiReader.Diablo
 {
     public static class IServiceCollectionExtension
     {
-        public static IServiceCollection AddBlizzardApiReaderDiablo(this IServiceCollection services)
+        public static IServiceCollection AddBlizzardApiReaderDiablo(this IServiceCollection services, HttpConfiguration httpConfiguration=null)
         {
-            services.AddBlizzardApiReader();
+            services.AddBlizzardApiReader(httpConfiguration);
             services.AddScoped<DiabloApi>();            
             return services;
         }
